// src/components/tables/client-table/client-table-types.ts
export type ClientTableTypes = {
  client_id: number;
  client_name: string;
  city_id: number;
  city_name: string;
  state_abb: string;
  country_id: string;
  industry_id: number;
  industry_name: string;
  company_size_id: string;
  company_desc: string;
  active: boolean;
  InsertDate: Date;
  ModifiedDate: Date;
};